Rename ping to system-info and integrate with status endpoint
- Rename /ping endpoint to /system-info for better semantics
- Rename PingHandler to SystemInfoHandler and related functions
- Rename Ping.go to SystemInfo.go
- Integrate sinfo -s call into StatusHandler for empty requests
- When StatusHandler receives empty pod list, return sinfo -s output
- Consolidate getSinfoSummary method in Status.go
- Remove duplicate imports and functions

This allows the /status endpoint to automatically provide system
information when called without specific pods, fulfilling the
requirement to show sinfo -s data for node annotations.
